refactor: reorganize headers and dynamic admin tab title
