+++ frontend/src/components/PageHeader.tsx
@@ -1,13 +1,23 @@
-import React from "react";
+import React from 'react';
 
-interface PageHeaderProps {
-    title: string;
-}
-
-const PageHeader: React.FC<PageHeaderProps> = ({ title }) => (
-    <header className="pb-2 mb-1">
-        <h1 className="text-xl sm:text-2xl font-semibold text-center sm:text-left">{title}</h1>
-    </header>
-);
+const PageHeader: React.FC = () => {
+    return (
+        <div
+            className="w-full bg-cover bg-center"
+            style={{ backgroundImage: "url('/header-background.png')" }}
+        >
+            <div className="bg-black/50">
+                <div className="mx-auto w-full max-w-5xl px-4 py-8 sm:px-8 sm:py-10">
+                    <h1 className="text-2xl font-bold tracking-tight text-white sm:text-3xl">
+                        Conference Registration
+                    </h1>
+                    <p className="mt-2 text-sm font-medium text-white/90 sm:text-base">
+                        Invitation-only access
+                    </p>
+                </div>
+            </div>
+        </div>
+    );
+};
 
 export default PageHeader;
